"Fixes und Linux-Anpassung"
git-svn-id: svn://svn.berlios.de/dsa-hl@1730 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/multimedia/src/VorbisFile.cpp
@@ -40,7 +40,7 @@ VorbisFile::VorbisFile(const SoundResourcePtr &res):
     mVorbisInfo = *ov_info(&mOggStream, -1);
     double time = ov_time_total(&mOggStream, -1);
     mTotalTime.sec = (long long int)time;
-    mTotalTime.nsec = (int)(time - mTotalTime.sec) * pow(10.0, 9.0);
+    mTotalTime.nsec = (int)(time - (double)mTotalTime.sec) * pow(10.0, 9.0);
 }
 
 VorbisFile::~VorbisFile()
